add fabfile

+++ raspberry-pi/camera-agent/main.py
@@ -8,7 +8,7 @@ import picamera
 
 import face_detect
 
-def take_photo():
+def take_photo(config):
   '''take photo by pi camera'''
   # code recipe from
   # https://picamera.readthedocs.org/en/release-0.7/recipes1.html
@@ -17,25 +17,36 @@ def take_photo():
     camera.start_preview()
     # Camera warm-up time
     time.sleep(2)
-    # use current time as filename to prevent duplicating files
-    filename = datetime.datetime.now().strftime('%Y%m%d-%H-%M-%S.jpg')
+    filepath = determine_photo_path(config)
     # TODO use log
-    print 'capture to %s' % filename
-    # TODO use capture dir setting
-    camera.capture(filename)
-    return filename
+    print 'capture to %s' % filepath
+    camera.capture(filepath)
+    return filepath
+
+def determine_photo_path(config):
+  # use current time as filename to prevent duplicating files
+  filename = datetime.datetime.now().strftime('%Y%m%d-%H-%M-%S.jpg')
+  capture_dir = config.get('Default', 'CaptureDir')
+  ensure_directory(capture_dir)
+  return os.path.join(capture_dir, filename)
+
+# TODO can be a utility method
+def ensure_directory(path):
+  if not os.path.isdir(path):
+    print 'directory [%s] not exist, create it' % path
+    os.mkdirs(path)
 
 def upload_photo(filename, config):
-  if config.getboolean('Default', 'DeletePhotoWithoutFace'):
+  if config.getboolean('Default', 'CountFace'):
     face_count = face_detect.count_face(filename)
-    if face_count == 0:
+    if config.getboolean('Default', 'DeletePhotoWithoutFace') and face_count == 0:
       # no face in photo, just delete this photo
       # TODO use log
       print 'no face, delete photo %s' % filename
       os.remove(filename)
       return
     else:
-      print '%d face detected, keep going' % face_count
+      print '%d face detected' % face_count
 
   print 'start uploading'
   
@@ -51,5 +62,5 @@ def load_config(config_filename):
 if __name__ == '__main__':
   # TODO separate config
   config = load_config('settings.ini')
-  filename = take_photo()
+  filename = take_photo(config)
   upload_photo(filename, config)
